Added some user restrictions

+++ blog/views.py
@@ -57,6 +57,10 @@ def detailed_blog_view(request, blog_id):
 @login_required
 def edit_post(request, post_id):
     """ edit blog post. """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('blogs'))
+
     post = get_object_or_404(blog, pk=post_id)
 
     if request.method == 'POST':
@@ -79,6 +83,10 @@ def edit_post(request, post_id):
 @login_required
 def delete_post(request, post_id):
     """ Deletes blog post. """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('blogs'))
+
     get_object_or_404(blog, pk=post_id).delete()
     messages.success(request, 'Your blog post has been successfully Deleted!.')
     return redirect(reverse('blogs'))
